+++ src/main/java/com/STC/StudyCode/Category/Dto/CategoryInfoDto.java
@@ -2,5 +2,5 @@ package com.STC.StudyCode.Category.Dto;
 
 public interface CategoryInfoDto {
     String getCategoryName();
-    Integer getCategoryIndex();
+    Long getCategoryIndex();
 }
